Normalize foresight_default_ttl_days to f64; migrate bench_beam to add_note_with_clock
Two follow-ups from the STEP1 mega-merge:

(1) WriteConfig had a typing inconsistency: foresight_default_ttl_days
was i64 while the new foresight_doa_threshold_days and
future_skew_threshold_days fields landed as f64. Normalize all three
to f64 so the convention is consistent. Updates the Duration computation
in write.rs to convert via milliseconds since chrono::Duration::days
takes i64.

(2) bench_beam.rs still called the deleted add_note_with_metadata.
Migrate to add_note_with_clock with an explicit ClockContext::at when
source_timestamp is Some, ::now() otherwise. This test is lance-feature
gated so it didn't run pre-migration anyway, but cleaning up the API
call site so future lance-feature work has correct API.

+++ crates/karta-core/src/config.rs
@@ -163,7 +163,7 @@ pub struct WriteConfig {
     /// Max evolutions before a note is flagged for consolidation instead.
     pub max_evolutions_per_note: usize,
     /// Default TTL in days for foresight signals when no explicit expiry is extracted.
-    pub foresight_default_ttl_days: i64,
+    pub foresight_default_ttl_days: f64,
     /// Whether to extract and store atomic facts during note ingestion.
     pub extract_atomic_facts: bool,
     /// Maximum number of atomic facts to extract per note.
@@ -193,7 +193,7 @@ impl Default for WriteConfig {
             similarity_threshold: 0.3,
             evolve_linked_notes: true,
             max_evolutions_per_note: 5,
-            foresight_default_ttl_days: 90,
+            foresight_default_ttl_days: 90.0,
             extract_atomic_facts: true,
             max_facts_per_note: 5,
             foresight_doa_threshold_days: 1.0,
